refactor(annotator): audio annotator components are in their own folder now
