+++ src/AudioAnnotator/AudioPlayer.js
